Discount as decimal, smart date picker UX, softer form style
- DiscountPercent: int → decimal(5,3) in model, VM, DbContext and snapshot
- Product.GetDiscountLabel() formats percent without trailing zeros
- Date fields: after picking start date end date auto-opens; focusing end without start redirects to start
- Promotion form: flat bg-light inputs, vr separators, switch toggle for IsActive, softer layout
- All three views (Home, Catalog, Detail) use GetDiscountLabel() for the badge text

+++ Algara.Data/Models/Promotion.cs
@@ -6,7 +6,7 @@
         public string Name { get; set; } = string.Empty;
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
-        public int DiscountPercent { get; set; }
+        public decimal DiscountPercent { get; set; }
         public bool IsActive { get; set; } = true;
         public DateTime CreatedAt { get; set; } = DateTime.Now;
